# internal/httpapi/openapi.yaml
openapi: 3.1.0
info:
  title: mango
  version: 0.1.0
  description: >-
    Mango's current HTTP API for durable Sessions, Deployments, self-hosted
    Environment Work, conditional Files and Memory APIs, Vaults, Webhooks, and
    Memory-backed Session Resources. Mango is an independent, self-hosted
    runtime licensed under Apache-2.0.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: /
    description: Current mango server
security:
  - BearerAuth: []
...
paths:
  /v1/agents:
    post:
      operationId: createAgent
      x-sdk-resource: agents
      x-sdk-method: create
      tags: [Agents]
      summary: Create an Agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCreateRequest'
      responses:
        '200':
          $ref: '#/components/responses/AgentResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    get:
      operationId: listAgents
      x-sdk-resource: agents
      x-sdk-method: list
      tags: [Agents]
      summary: List Agents
      parameters:
        - $ref: '#/components/parameters/AgentListLimit'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/IncludeArchived'
        - $ref: '#/components/parameters/CreatedAtGte'
        - $ref: '#/components/parameters/CreatedAtLte'
      responses:
        '200':
          $ref: '#/components/responses/AgentListResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
  /v1/agents/{agent_id}:
    parameters:
      - $ref: '#/components/parameters/AgentID'
    get:
      operationId: getAgent
      x-sdk-resource: agents
      x-sdk-method: retrieve
      tags: [Agents]
      summary: Get an Agent
      responses:
        '200':
          $ref: '#/components/responses/AgentResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      operationId: updateAgent
      x-sdk-resource: agents
      x-sdk-method: update
      tags: [Agents]
      summary: Update an Agent and create a new version when fields change
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentUpdateRequest'
      responses:
        '200':
          $ref: '#/components/responses/AgentResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
  /v1/agents/{agent_id}/versions:
    parameters:
      - $ref: '#/components/parameters/AgentID'
    get:
      operationId: listAgentVersions
      x-sdk-resource: agents.versions
      x-sdk-method: list
      tags: [Agents]
      summary: List Agent versions
      parameters:
        - $ref: '#/components/parameters/AgentListLimit'
        - $ref: '#/components/parameters/Page'
      responses:
        '200':
          $ref: '#/components/responses/AgentListResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/agents/{agent_id}/archive:
    parameters:
      - $ref: '#/components/parameters/AgentID'
    post:
      operationId: archiveAgent
      x-sdk-resource: agents
      x-sdk-method: archive
      tags: [Agents]
      summary: Archive an Agent
      responses:
        '200':
          $ref: '#/components/responses/AgentResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/environments:
    post:
      operationId: createEnvironment
      x-sdk-resource: environments
      x-sdk-method: create
      tags: [Environments]
      summary: Create an Environment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvironmentCreateRequest'
      responses:
        '200':
          $ref: '#/components/responses/EnvironmentResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    get:
      operationId: listEnvironments
      x-sdk-resource: environments
      x-sdk-method: list
      tags: [Environments]
      summary: List Environments
      parameters:
        - $ref: '#/components/parameters/EnvironmentListLimit'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/IncludeArchived'
      responses:
        '200':
          $ref: '#/components/responses/EnvironmentListResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
  /v1/environments/{environment_id}:
    parameters:
      - $ref: '#/components/parameters/EnvironmentID'
    get:
      operationId: getEnvironment
      x-sdk-resource: environments
      x-sdk-method: retrieve
      tags: [Environments]
      summary: Get an Environment
      responses:
        '200':
          $ref: '#/components/responses/EnvironmentResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      operationId: updateEnvironment
      x-sdk-resource: environments
      x-sdk-method: update
      tags: [Environments]
      summary: Update an Environment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvironmentUpdateRequest'
      responses:
        '200':
          $ref: '#/components/responses/EnvironmentResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    delete:
      operationId: deleteEnvironment
      x-sdk-resource: environments
      x-sdk-method: delete
      tags: [Environments]
      summary: Delete an unreferenced Environment
      responses:
        '200':
          $ref: '#/components/responses/EnvironmentDeletedResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/environments/{environment_id}/archive:
    parameters:
      - $ref: '#/components/parameters/EnvironmentID'
    post:
      operationId: archiveEnvironment
      x-sdk-resource: environments
      x-sdk-method: archive
      tags: [Environments]
      summary: Archive an Environment
      responses:
        '200':
          $ref: '#/components/responses/EnvironmentResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/environments/{environment_id}/work:
    parameters:
      - $ref: '#/components/parameters/EnvironmentID'
    get:
      operationId: listEnvironmentWork
      x-sdk-resource: environments.work
      x-sdk-method: list
      tags: [Environment Work]
      summary: List self-hosted Environment Work items
      parameters:
        - $ref: '#/components/parameters/EnvironmentWorkListLimit'
        - $ref: '#/components/parameters/Page'
      responses:
        '200': {$ref: '#/components/responses/EnvironmentWorkListResponse'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
  /v1/environments/{environment_id}/work/poll:
    parameters:
      - $ref: '#/components/parameters/EnvironmentID'
    get:
      operationId: pollEnvironmentWork
      x-sdk-resource: environments.work
      x-sdk-method: poll
      tags: [Environment Work]
      summary: Poll and tentatively claim the oldest available Work item
      parameters:
        - $ref: '#/components/parameters/EnvironmentWorkBlockMs'
        - $ref: '#/components/parameters/EnvironmentWorkReclaimMs'
        - $ref: '#/components/parameters/EnvironmentWorkerID'
      responses:
        '200': {$ref: '#/components/responses/EnvironmentWorkPollResponse'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
  /v1/environments/{environment_id}/work/stats:
    parameters:
      - $ref: '#/components/parameters/EnvironmentID'
    get:
      operationId: getEnvironmentWorkStats
      x-sdk-resource: environments.work
      x-sdk-method: stats
      tags: [Environment Work]
      summary: Get Environment Work queue statistics
      responses:
        '200': {$ref: '#/components/responses/EnvironmentWorkStatsResponse'}
        '404': {$ref: '#/components/responses/NotFound'}
  /v1/environments/{environment_id}/work/{work_id}:
    parameters:
      - $ref: '#/components/parameters/EnvironmentID'
      - $ref: '#/components/parameters/EnvironmentWorkID'
    get:
      operationId: getEnvironmentWork
      x-sdk-resource: environments.work
      x-sdk-method: retrieve
      tags: [Environment Work]
      summary: Get an Environment Work item
      responses:
        '200': {$ref: '#/components/responses/EnvironmentWorkResponse'}
        '404': {$ref: '#/components/responses/NotFound'}
    post:
      operationId: updateEnvironmentWork
      x-sdk-resource: environments.work
      x-sdk-method: update
      tags: [Environment Work]
      summary: Merge metadata into an Environment Work item
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/EnvironmentWorkUpdateRequest'}
      responses:
        '200': {$ref: '#/components/responses/EnvironmentWorkResponse'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
  /v1/environments/{environment_id}/work/{work_id}/ack:
    parameters:
      - $ref: '#/components/parameters/EnvironmentID'
      - $ref: '#/components/parameters/EnvironmentWorkID'
    post:
      operationId: acknowledgeEnvironmentWork
      x-sdk-resource: environments.work
      x-sdk-method: ack
      tags: [Environment Work]
      summary: Acknowledge a polled Work item and transition it to starting
      responses:
        '200': {$ref: '#/components/responses/EnvironmentWorkResponse'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /v1/environments/{environment_id}/work/{work_id}/heartbeat:
    parameters:
      - $ref: '#/components/parameters/EnvironmentID'
      - $ref: '#/components/parameters/EnvironmentWorkID'
    post:
      operationId: heartbeatEnvironmentWork
      x-sdk-resource: environments.work
      x-sdk-method: heartbeat
      tags: [Environment Work]
      summary: Record a conditional heartbeat and extend the Work lease
      parameters:
        - $ref: '#/components/parameters/EnvironmentWorkExpectedHeartbeat'
        - $ref: '#/components/parameters/EnvironmentWorkDesiredTTL'
      responses:
        '200': {$ref: '#/components/responses/EnvironmentWorkHeartbeatResponse'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
  /v1/environments/{environment_id}/work/{work_id}/fail:
    parameters:
      - $ref: '#/components/parameters/EnvironmentID'
      - $ref: '#/components/parameters/EnvironmentWorkID'
    post:
      operationId: failEnvironmentWork
      x-sdk-resource: environments.work
      x-sdk-method: fail
      tags: [Environment Work]
      summary: Terminate a Session whose immutable worker inputs are invalid
      description: >-
        This terminal operation is reserved for permanent preparation failures.
        A worker experiencing a transient retrieval failure leaves the Work
        active so lease expiry and queue reclaim can retry it.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/EnvironmentWorkFailureRequest'}
      responses:
        '204':
          description: Work and Session failure committed; no response body
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
  /v1/environments/{environment_id}/work/{work_id}/stop:
    parameters:
      - $ref: '#/components/parameters/EnvironmentID'
      - $ref: '#/components/parameters/EnvironmentWorkID'
    post:
      operationId: stopEnvironmentWork
      x-sdk-resource: environments.work
      x-sdk-method: stop
      tags: [Environment Work]
      summary: Request graceful or forced Work shutdown
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/EnvironmentWorkStopRequest'}
      responses:
        '204':
          description: Work stop accepted; no response body
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /v1/files:
    post:
      operationId: uploadFile
      x-sdk-resource: files
      x-sdk-method: upload
      tags: [Files]
      summary: Upload a File
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FileUploadRequest'
      responses:
        '200':
          $ref: '#/components/responses/FileResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    get:
      operationId: listFiles
      x-sdk-resource: files
      x-sdk-method: list
      tags: [Files]
      summary: List Files
      parameters:
        - $ref: '#/components/parameters/FileListLimit'
        - $ref: '#/components/parameters/AfterID'
        - $ref: '#/components/parameters/BeforeID'
        - $ref: '#/components/parameters/ScopeID'
      responses:
        '200':
          $ref: '#/components/responses/FileListResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
  /v1/files/{file_id}:
    parameters:
      - $ref: '#/components/parameters/FileID'
    get:
      operationId: getFileMetadata
      x-sdk-resource: files
      x-sdk-method: retrieve
      tags: [Files]
      summary: Get File metadata
      responses:
        '200':
          $ref: '#/components/responses/FileResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    delete:
      operationId: deleteFile
      x-sdk-resource: files
      x-sdk-method: delete
      tags: [Files]
      summary: Delete a File
      responses:
        '200':
          $ref: '#/components/responses/FileDeletedResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
  /v1/files/{file_id}/content:
    parameters:
      - $ref: '#/components/parameters/FileID'
    get:
      operationId: downloadFile
      x-sdk-resource: files
      x-sdk-method: download
      tags: [Files]
      summary: Download a File marked as downloadable
      responses:
        '200':
          $ref: '#/components/responses/FileContentResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
  /v1/skills:
    post:
      operationId: createSkill
      x-sdk-resource: skills
      x-sdk-method: create
      tags: [Skills]
      summary: Create a custom Skill and its first immutable Version
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SkillUploadRequest'
      responses:
        '200':
          $ref: '#/components/responses/SkillResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    get:
      operationId: listSkills
      x-sdk-resource: skills
      x-sdk-method: list
      tags: [Skills]
      summary: List Skills
      parameters:
        - $ref: '#/components/parameters/SkillListLimit'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/SkillSource'
      responses:
        '200':
          $ref: '#/components/responses/SkillListResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
  /v1/skills/{skill_id}:
    parameters:
      - $ref: '#/components/parameters/SkillID'
    get:
      operationId: getSkill
      x-sdk-resource: skills
      x-sdk-method: retrieve
      tags: [Skills]
      summary: Get a Skill
      responses:
        '200':
          $ref: '#/components/responses/SkillResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    delete:
      operationId: deleteSkill
      x-sdk-resource: skills
      x-sdk-method: delete
      tags: [Skills]
      summary: Delete a Skill after all of its Versions have been deleted
      responses:
        '200':
          $ref: '#/components/responses/SkillDeletedResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
...
  /v1/skills/{skill_id}/versions/{version}:
    parameters:
      - $ref: '#/components/parameters/SkillID'
      - $ref: '#/components/parameters/SkillVersion'
    get:
      operationId: getSkillVersion
      x-sdk-resource: skills.versions
      x-sdk-method: retrieve
      tags: [Skills]
      summary: Get a Skill Version
      responses:
        '200':
          $ref: '#/components/responses/SkillVersionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    delete:
      operationId: deleteSkillVersion
      x-sdk-resource: skills.versions
      x-sdk-method: delete
      tags: [Skills]
      summary: Delete a Skill Version and its archived contents
      responses:
        '200':
          $ref: '#/components/responses/SkillVersionDeletedResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >-
        Workspace API key, or the per-Work sessions_token on the limited
        self-hosted execution routes authorized for that claim. The latter can
        read attached Memory Stores and mutate only read_write attachments while
        its Work lease remains live.
...
  responses:
...
    AgentResponse:
      description: Agent resource
      headers:
        request-id:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Agent'
...
  schemas:
...
    AgentCreateRequest:
      type: object
      additionalProperties: false
      required: [name, model]
      properties:
        name:
          type: string
          minLength: 1
        model:
          $ref: '#/components/schemas/ModelInput'
        system:
          type: [string, 'null']
        description:
          type: [string, 'null']
        tools:
          type: array
          maxItems: 128
          items:
            $ref: '#/components/schemas/AgentTool'
        mcp_servers:
          type: array
          maxItems: 20
          items:
            $ref: '#/components/schemas/MCPServer'
        skills:
          type: array
          maxItems: 500
          items:
            $ref: '#/components/schemas/SkillReferenceInput'
        multiagent:
          $ref: '#/components/schemas/MultiagentInput'
        metadata:
          $ref: '#/components/schemas/Metadata'
...